+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Revert to `DADF5` class if no `Result` class importable within `readers.read_HDF5_file`.
 - Change expected format of `microstructure_image` argument in function `volume_element_from_2D_microstructure`.
+- Explicitly track "P" constant in `orientations` dict.
+- Add `euler_degrees` boolean dict item to `orientations` dict.
 
 ## [0.2.7] - 2020.01.11